Debug test in linux
